Add tests demonstrating inconsistent attribute definition overwrites
When passing attribute definitions to create()/instance() it is unclear from the documentation how conflicts between the passed definitions and definitions created by calls to define() should be resolved.

testDefineWithReplacementGeneratorsOverwrite() demonstrates what I intuit to be expected behavior: the definitions passed to create() overwrite previous definitions.

testGroupDefineWithReplacementGeneratorsOverwrite() demonstrates behavior that conflicts with the above. The definition defined by the group factory overwrites the definitions passed in to create().