Adds support for setting model attributes through setters
Some model attributes needs to be set through setter methods, but
FactoryMuffin only sets the attributes directly. This fix checks for the
presence of a setter method on the model and calls it instead. Ideally the
model itself would check for the presence of the setter but this is not
always the case. I also included a camelize function. It could possibly be
refactored.